requirements.txt
# Core Data Science Libraries
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Machine Learning Libraries
scikit-learn==1.3.0
xgboost==1.7.6
imbalanced-learn==0.11.0

# Deep Learning (Optional)
tensorflow==2.13.0
keras==2.13.1

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Web Application
streamlit==1.27.0
streamlit-option-menu==0.3.6

# Jupyter Notebooks
jupyter==1.0.0
ipykernel==6.25.0

# Configuration and Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7

# Model Persistence
joblib==1.3.2
pickle-mixin==1.0.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Code Quality
black==23.7.0
flake8==6.0.0
isort==5.12.0

# Additional Utilities
tqdm==4.65.0
requests==2.31.0
